tema3/adapterMQTT-DB/requirements.txt
paho-mqtt
influxdb
pytz
